Fixed: arch fixed, hardcode jwt key, some test

+++ internal/core/ports/users_repository.go
@@ -5,8 +5,8 @@ import (
 )
 
 type UsersRepository interface {
-	UserByID(id uint) (*domain.User, error)
+	UserByID(id int64) (*domain.User, error)
 	AddUser(user *domain.User) (*domain.User, error)
 	UserByLogin(login string) (*domain.User, error)
-	DeleteUser(id uint) error
+	DeleteUser(id int64) error
 }
